requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
fake-useragent==1.5.1
idna==3.10
importlib_resources==6.4.5
pillow==11.0.0
prettytable==3.12.0
pyperclip==1.9.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
requests==2.32.3
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.21.0